google_places/source_code/requirements.txt
requests==2.26.0
google-cloud-storage==1.43.0